+++ effectful-core/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Make `withEffToIO` take an explicit unlifting strategy for the sake of
   consistency with unlifting functions from `Effectful.Dispatch.Dynamic` and
   easier to understand API.
+* Add support for turning an effect handler into an effectful operation via the
+  `Reified` effect.
 
 # effectful-core-2.2.2.2 (2023-03-13)
 * Allow `inject` to turn a monomorphic effect stack into a polymorphic one.